planettool: Optimizations - less sign extensions, float precision conversions, and sqrts.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk/tools/planettool@3211 127b21dd-08f5-0310-b4b7-95ae10353056

+++ FloatPixMap/FPMBasics.h
@@ -49,8 +49,8 @@ FPM_BEGIN_EXTERN_C
 
 
 typedef float		FPMComponent;
-typedef intmax_t	FPMCoordinate;
-typedef uintmax_t	FPMDimension;
+typedef int32_t		FPMCoordinate;
+typedef uint32_t	FPMDimension;
 
 
 // FPMColor: an RGBA colour value.
